Moves "is_io" function from "functions.py" to a new file called "instances.py"
